backend/requirements.txt
fastapi
uvicorn
python-multipart==0.0.9
requests==2.31.0
google-generativeai
python-dotenv
imageio-ffmpeg
assemblyai
